backend/requirements.txt
Flask
gunicorn
celery[redis]
pillow